+++ spec/byciclist_spec.rb
@@ -1,6 +1,7 @@
 require 'sinatra'
 require 'sinatra/bicyclist'
 
+set :environment, :test
 set :routes_to_cycle_through, {
     one: [:foo],
     two: [:bar, :baz]
@@ -14,7 +15,6 @@ RSpec.describe Sinatra::Application do
 
     it { expect(response.status).to eq(302) }
     it { expect(response.headers).to include('Location' => 'http://example.org/_cycle/one') }
-
 
     context 'two' do
       before { get '/_cycle/two' }